+++ scripts/World/Heal.cs
@@ -4,7 +4,7 @@ using System;
 public partial class Heal : Node
 {
 	[Export] public Item _healItem;
-	[Export] public float _healAmount;
+	[Export] public int _healAmount;
 
 	[Export] public string _healMessage;
